+++ src/index.ts
@@ -1,7 +1,9 @@
 export { LiveCodeBlock } from "./LiveCodeBlock";
+export { LiveCodeDocsBlock } from "./LiveCodeDocsBlock";
 export type {
   LiveCodeBlockProps,
   LiveCodeMode,
   LiveCodeScope,
   LiveCodeTheme
 } from "./LiveCodeBlock";
+export type { LiveCodeDocsBlockProps } from "./LiveCodeDocsBlock";